Skip non-existent mailing lists from the task queue.
It is possible in certain situations that a mailing list is deleted
after a task for it is queued in the async task queue. This results in a
MailingList.DoesNotExist being raised, whereupon the task is put back
into the async task queue to be re-executed again at a later time.

Add an exception handler, like other tasks in hyperkitty.tasks have, so
that running a task for a non-existent mailing list fails gracefully.