Skip listing of private archives in the search.
This esssentially makes lists with private archives un-searchable
in the search. The search currently does not filter based on the
VHOST or based on the membership of the currently logged in user.

Also, even though we use `request.user.is_superuser`, the fact that
the call is made via JS with no django proper django context, there
is no django token to authenticate the currently logged in user.

Fixes #237