Do not evaluate the entire query string to get count.
Using `if` on a query string evaluates the entire queryset and fetches all the
rows for the threads from the database. The larger the list, the slower this
gets. So, we instead use .count() and then let pagination load only the
required threads from the database.

This shouls speed up the /latest endpoints for mailing list significantly.